Create file_converter.rb
